Added lots of debugging output while I attempt to locate the source of audio scratchiness, also simplified manner of dealing with wraparound within the AudioQueue wrapper.

+++ OSBindings/Mac/Clock Signal/Wrappers/AudioQueue.h
@@ -10,6 +10,6 @@
 
 @interface AudioQueue : NSObject
 
-- (void)enqueueAudioBuffer:(const int16_t *)buffer numberOfSamples:(unsigned int)lengthInSamples;
+- (void)enqueueAudioBuffer:(const int16_t *)buffer numberOfSamples:(size_t)lengthInSamples;
 
 @end
